update clustering by same taxa
