+++ bims/urls.py
@@ -14,7 +14,7 @@ from bims.api_views.location_type import (
 )
 from bims.api_views.taxon import TaxonDetail
 from bims.api_views.cluster import ClusterList
-from bims.api_views.cluster_module import ClusterModuleList
+from bims.api_views.cluster_taxa import ClusterTaxaList
 from bims.views.links import LinksCategoryView
 from bims.api_views.search import SearchObjects
 from bims.views.activate_user import activate_user
@@ -30,8 +30,8 @@ api_urls = [
         TaxonDetail.as_view()),
     url(r'^api/cluster/(?P<administrative_level>\w+)/$',
         ClusterList.as_view()),
-    url(r'^api/cluster-module/$',
-        ClusterModuleList.as_view()),
+    url(r'^api/cluster-data/taxa/(?P<pk>[0-9]+)/$',
+        ClusterTaxaList.as_view()),
     url(r'^api/search/(?P<query_value>\w+)/$',
         SearchObjects.as_view(), name='search-api'),
 ]
